# docs/swagger.yaml
basePath: /
definitions:
  apikey.Key:
    properties:
      created_at:
        type: string
      id:
        type: integer
      key_hint:
        description: Masked representation for display (e.g. "sk-or-v1-...1234")
        type: string
      provider:
        allOf:
        - $ref: '#/definitions/types.Provider'
        description: ProviderOpenRouter, ProviderOpenAI, ProviderAnthropic, ProviderGoogle
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  apikey.RegisterAPIKeyRequest:
    properties:
      api_key:
        maxLength: 500
        type: string
      provider:
        allOf:
        - $ref: '#/definitions/types.Provider'
        description: '"openrouter", "openai", "anthropic", "google"'
    required:
    - api_key
    - provider
    type: object
  chat.Chat:
    properties:
      created_at:
        type: string
      id:
        type: integer
      messages:
        items:
          $ref: '#/definitions/chat.Message'
        type: array
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  chat.CreateChatRequest:
    properties:
      title:
        maxLength: 255
        type: string
    type: object
  chat.Message:
    properties:
      chat_id:
        type: integer
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      role:
        allOf:
        - $ref: '#/definitions/chat.Role'
        description: '"user", "assistant", "system"'
    type: object
  chat.Role:
    enum:
    - user
    - assistant
    - system
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleAssistant
    - RoleSystem
  chat.SendMessageRequest:
    properties:
      content:
        maxLength: 30000
        type: string
      model:
        maxLength: 100
        type: string
      provider:
        $ref: '#/definitions/types.Provider'
    required:
    - content
    - model
    - provider
    type: object
  profile.Profile:
    properties:
      created_at:
        type: string
      system_prompt:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  profile.UpdateSystemPromptRequest:
    properties:
      system_prompt:
        maxLength: 10000
        type: string
    type: object
  types.Provider:
    enum:
    - openrouter
    - openai
    - anthropic
    - google
    type: string
    x-enum-varnames:
    - ProviderOpenRouter
    - ProviderOpenAI
    - ProviderAnthropic
    - ProviderGoogle
host: localhost:8080
info:
  contact: {}
  description: Go + Gin + Clerk + GORM による AI チャット基盤 API
  title: Study Gin Clerk & AI Chat API
  version: "1.0"
paths:
  /api/v1/ai/models:
    get:
      description: 認証済みユーザーが登録した API キーに基づき、プロバイダから動的に取得したモデル一覧を返却します。15分間のインメモリキャッシュ付きです。
      parameters:
      - description: キャッシュをバイパスして強制再取得するか
        in: query
        name: refresh
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未認証
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: サーバーエラー
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 利用可能モデル一覧取得 (動的取得 & キャッシュ)
      tags:
      - ai_models
  /api/v1/chats:
    get:
      description: 認証済みユーザーが作成したチャット一覧を取得します
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/chat.Chat'
              type: array
            type: object
        "401":
          description: 未認証
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 認可エラー・トークン不正
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: サーバーエラー
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 自分のチャット一覧取得
      tags:
      - chats
    post:
      consumes:
      - application/json
      description: 新しい会話セッションを作成します（モデル指定は不要でメッセージ送信時に選択します）
      parameters:
      - description: チャット設定 (タイトル, 最大255文字)
        in: body
        name: request
        schema:
          $ref: '#/definitions/chat.CreateChatRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/chat.Chat'
        "400":
          description: 不正なリクエスト
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未認証
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 認可エラー・トークン不正
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: サーバーエラー
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: チャット新規作成
      tags:
      - chats
  /api/v1/chats/{id}:
    get:
      description: 指定されたIDのチャット詳細と過去メッセージ履歴を取得します
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/chat.Chat'
        "400":
          description: 不正なID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未認証
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 認可エラー・トークン不正
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: チャットが見つからない
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: チャット詳細・メッセージ履歴取得
      tags:
      - chats
  /api/v1/chats/{id}/messages:
    post:
      consumes:
      - application/json
      description: ユーザーメッセージを送信し、指定されたプロバイダ・モデルを使ってAI返答を一括生成して保存します（一括JSON）
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: integer
      - description: メッセージ内容・プロバイダ・モデル
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/chat.SendMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/chat.Message'
            type: object
        "400":
          description: 不正なリクエスト
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未認証
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: チャットが見つからない
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: サーバーエラー
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: AIプロバイダ通信エラー
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: メッセージ送信＆AI返答一括生成
      tags:
      - chats
  /api/v1/chats/{id}/messages/stream:
    post:
      consumes:
      - application/json
      description: ユーザーメッセージを送信し、指定されたプロバイダ・モデルでAI返答を Server-Sent Events (SSE) で1文字ずつリアルタイム受信します。完了時に全文が自動保存されます。
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: integer
      - description: メッセージ内容・プロバイダ・モデル
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/chat.SendMessageRequest'
      produces:
      - text/event-stream
      responses:
        "200":
          description: text/event-stream
          schema:
            type: string
        "400":
          description: 不正なリクエスト
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未認証
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: チャットが見つからない
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: サーバーエラー
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: AIプロバイダ通信エラー
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: メッセージ送信＆AI返答リアルタイムストリーミング (SSE)
      tags:
      - chats
  /api/v1/me:
    get:
      description: 認証済みユーザーのプロファイル情報およびシステムプロンプトを取得します
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profile.Profile'
        "401":
          description: 未認証
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 認可エラー・トークン不正
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: サーバーエラー
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 自分のプロファイル取得
      tags:
      - user_profiles
  /api/v1/me/api-keys:
    get:
      description: ユーザーが登録した各社 API キーの一覧（マスクされた key_hint のみ）を取得します。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/apikey.Key'
              type: array
            type: object
        "401":
          description: 未認証
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: サーバーエラー
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 登録済み API キー一覧
      tags:
      - user_api_keys
    post:
      consumes:
      - application/json
      description: ユーザー独自の各社 AI API キーを登録します。登録時にプロバイダへの疎通確認を行い、AES-256-GCM で暗号化して安全に保存します。
      parameters:
      - description: API キー情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apikey.RegisterAPIKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 登録成功 (マスクされた key_hint を返却)
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 検証失敗または不正なリクエスト
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未認証
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: サーバーエラー
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: API キー登録・更新
      tags:
      - user_api_keys
  /api/v1/me/api-keys/{provider}:
    delete:
      description: 指定されたプロバイダの API キーを削除し、モデルキャッシュをパージします。
      parameters:
      - description: Provider Name (e.g. openrouter)
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 削除成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 不正なリクエスト
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未認証
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: キーが見つからない
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: サーバーエラー
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 登録済み API キー削除
      tags:
      - user_api_keys
  /api/v1/me/system-prompt:
    put:
      consumes:
      - application/json
      description: ユーザー共通のデフォルトシステムプロンプトを更新・保存します
      parameters:
      - description: システムプロンプト設定 (最大10000文字)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/profile.UpdateSystemPromptRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profile.Profile'
        "400":
          description: 不正なリクエスト
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未認証
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: サーバーエラー
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: システムプロンプト更新
      tags:
      - user_profiles
  /health:
    get:
      description: サーバーおよびデータベースの稼働状態を確認します
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: データベース接続異常
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ヘルスチェック (DB疎通確認)
      tags:
      - health
securityDefinitions:
  BearerAuth:
    description: Clerk から取得した JWT トークン (Bearer <TOKEN>)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"